Add commands API
